Add U+1F5B3 and U+1F5B5 (former from even older U+1F5A5, see below).
U+1F5B5 is derived from U+1F5A5, whereas U+1F5B3 is based on U+1F5B5 and
U+1F5AE from this set plus the U+1F5A5 from the older EmojiOne 2.0 (as
opposed to 2.2, which this set forked from), with logotype removed. This
is an older, as opposed to newer, version, so licence issues ought not be
a problem.